requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.25
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing-extensions==4.13.2
whitenoise==6.7.0
